// src/models/AlgorithmicPatterns.model.ts
export interface AlgorithmicPatterns {
  isBinarySearch: boolean;
  isDivideByHalf: boolean;
  isSorting: boolean;
  isFactorial: boolean;
  isHashAccess: boolean;
  isDirectAccess: boolean;
  isPermutations: boolean;
  isPowerSet: boolean;
  isBuiltinSort: boolean;
  hasExponentialLoop: boolean;
}